synvqa/requirements.txt
pyyaml>=6.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
scikit-learn>=1.3
anthropic>=0.34
openai>=1.40
tavily-python>=0.3
pillow>=10.0
pytesseract>=0.3.10
# Optional (for local target_vlm / faithfulness_vlm / t2i inference):
# transformers
# torch
# diffusers
# accelerate
